Re-do agg partials
