+++ stan/math/rev/fun/aggregate_partial.hpp
@@ -1,0 +1,12 @@
+#ifndef STAN_MATH_REV_AGGREGATE_PARTIAL_HPP
+#define STAN_MATH_REV_AGGREGATE_PARTIAL_HPP
+
+#include <stan/math/rev/meta.hpp>
+
+namespace stan {
+namespace math {
+
+
+}  // namespace math
+}  // namespace stan
+#endif
